+++ src/plugins/gui/providers/bityProvider.ts
@@ -84,7 +84,8 @@ const asBityCurrency = asObject({
 })
 const asBityCurrencyResponse = asObject({ currencies: asArray(asBityCurrency) })
 
-const asBityErrorResponse = asObject({ errors: asArray(asObject({ code: asString, message: asString })) })
+const asBityError = asObject({ code: asString, message: asString })
+const asBityErrorResponse = asObject({ errors: asArray(asBityError) })
 
 // const asCurrencyAmount = asObject({
 //   amount: asString,
@@ -219,6 +220,14 @@ const asBityApproveQuoteResponse = asEither(asBityBuyApproveQuoteResponse, asBit
 
 type BityApproveQuoteResponse = ReturnType<typeof asBityApproveQuoteResponse>
 
+class BityError extends Error {
+  code: string
+  constructor(message: string, code: string) {
+    super(message)
+    this.code = code
+  }
+}
+
 const fetchBityQuote = async (bodyData: BityQuoteRequest) => {
   const request = {
     method: 'POST',
@@ -242,7 +251,7 @@ const fetchBityQuote = async (bodyData: BityQuoteRequest) => {
       // is a required FiatProviderQuoteError param.
       throw new Error('Bity: Unable to fetch quote: ' + (await result.text()))
     }
-    if (bityErrorRes.errors.some((bityError: { code: string; message: string }) => bityError.code === 'amount_too_large')) {
+    if (bityErrorRes.errors.some(bityError => bityError.code === 'amount_too_large')) {
       throw new FiatProviderError({ providerId, errorType: 'overLimit' })
     }
   }
@@ -271,7 +280,7 @@ const approveBityQuote = async (
 
   if (orderRes.status !== 201) {
     const errorData = await orderRes.json()
-    throw new Error(errorData.errors[0].code + ' ' + errorData.errors[0].message)
+    throw new BityError(errorData.errors[0].message, errorData.errors[0].code)
   }
   // "location": "https://...bity.com/v2/orders/[orderid]"
   const locationHeader = orderRes.headers.get('Location')
@@ -557,6 +566,12 @@ export const bityProvider: FiatProviderFactory = {
                   // their personal info depending on what was wrong with the
                   // order, i.e. invalid bank or address info.
                   console.error('Bity order error: ', e)
+
+                  const bityError = asMaybe(asBityError)(e)
+                  if (bityError?.code === 'exceeds_quota') {
+                    await showUi.showError(sprintf(lstrings.error_kyc_required_s, bityError.message))
+                    return
+                  }
                   await showUi.showError(lstrings.error_unexpected_title)
                 }